+++ skimage/filters/diffusion_linear.py
